Change initializeTime to advance startTime_ when possible. The startTime_ is set as follows:
If there is demand...
    If there are inventory reports
        Set to earliest of 14 days before earliest demand or earliest
        inventory report.
    If there are no inventory reports
        Set to 14 days before earliest demand
If there is no demand...
    If there are inventory reports
        Set to earliest inventory report
    If there are no inventory reports
        Set to today

The firstDayOfDemand_ variable is set as follows:
If there is demand...
    Set to 14 days before earliest demand
If there is no demand...
    Set to today + 14 days

The start time is always set according to the above algorithm. It is
next to impossible to get the case where there is no demand are there
are no inventory reports because no inventories are created if there
is no demand.