Fix asset transfer loop.
There is an AssetTransfer to self (getAsset() and getAssignee() are
identical) in the logplan created to supply the "SELF" Relationship.
resendAssetTransfers was being called in response to a change of the
"Self" Organization and doing publishChange on all the AssetTransfers
from Self to wherever including the AssetTransfer to self. When the
resultant AssetAssignment arrives back in the cluster, the Assignee
asset is publishChanged, but this is the same Organization asset that
kicked the whole thing off in the first place.