Accept two new input parameters - OffsetDays and OriginFile. OffsetDays sets the start date of the Task to be the Oplan C day plus the value of the parameter.
OriginFile specifies an XML file containing a GeolocLocation to be used
as the FROM location of the Task. If the value of the parameter is "HOME"
or not specified, the home location of the Organization is used. If the home
location is not used, a PREPO prepositional phrase is added to the Task.