Fix for bug 3505. From Aaron's summary on the bug report - The OplanWatcherLP rebuilds the self org location schedule from scratch any time an OrgActivity changes. If a new OrgActivity is added, it adds that schedule element to the location schedule. It then publish Changes the OrgAsset with an anonymous Change report. This in turn kicks the OrgReportPlugin, AssetTransferLP, and then PropogationPlugin and GenericTablePlugin.
It seems the OplanWatcherLP could avoid publish changing the OrgAsset if it
first noticed wither the location schedule was really changing due to the new
OrgActivity (maybe it always is? Doesn't seem like it.)

Proposed fix:

1) Create new LocalPG.LocalPGChangeReport
2) OplanWatcherLP puts this ChangeReport on when it is only changing the
LocationSchedulePG
3) AssetReportPlugin does not resendAssetTransfers if all the ChangeReports
are either this LocalPG.LocalPGChangeReport or a
RelationshipSchedule.RelationshipScheduleChangeReport

The result is that a new OrgActivity does not require re-sending the local
Org to everyone you report to, and does not require re-sending all of your
Policy's (from the PropogationPlugin).